refactor!: harmonize project structure with kedro-azureml-pipeline
Consolidate config models into single config/models.py, split CLI into
a package (cli/__init__.py, cli/commands.py, cli/functions.py), extract
shared constants into constants.py, and add CliContext dataclass.

- Drop Pydantic v1 compat layer (require pydantic>=2.6.4)
- Bump kedro>=1.0.0 and kedro-datasets>=1.0.0, remove version branching
- Add CONFIG_TEMPLATE_YAML and validated _CONFIG_TEMPLATE instance
- Add Hypothesis property-based tests for config models
- Rename test files to mirror source modules (git mv)
- Raise interrogate fail-under to 100 and add missing docstrings
- Move migrate-existing-project to tutorials, rename mlflow/troubleshoot
  docs, create data-flow/hook-lifecycle explanation pages and datasets
  reference page, update mkdocs.yml nav and all internal links
- Fix executor validator bug when executors is None

BREAKING CHANGE: requires pydantic>=2.6.4, kedro>=1.0.0,
kedro-datasets>=1.0.0. CLI entry point changed to
kedro_dagster.cli.commands:commands. Config imports moved from
kedro_dagster.config.<submodule> to kedro_dagster.config.
