+++ docs/pages/explanation/index.md
@@ -1,0 +1,6 @@
+# Explanation
+
+Background and context to deepen your understanding of Kedro-Dagster's design and behavior.
+
+- **[Concepts](concepts.md)** - The asset-first alignment between Kedro and Dagster, key features, and what the integration provides for each framework's users.
+- **[Architecture](architecture.md)** - How Kedro projects are translated into Dagster code locations: catalog, node, pipeline, and hook mapping.
