+++ docs/pages/how-to/configure-logging.md
@@ -117,3 +117,9 @@ loggers:
 ```
 
 See the [Example Project](../tutorials/example-project.md#custom-logging-integration) for a complete logging setup across environments.
+
+## See also
+
+- [Configuration Reference](../reference/configuration.md#loggers) - `dagster.yml` logger field tables
+- [Troubleshooting](troubleshoot.md#logs-not-appearing-in-dagster-ui) - common logging issues
+- [Architecture](../explanation/architecture.md) - how Kedro logging integrates with Dagster
