+++ docs/pages/how-to/index.md
@@ -1,0 +1,26 @@
+# How-to Guides
+
+Practical directions for common tasks. Each guide addresses a specific goal and assumes you already have a working Kedro-Dagster setup.
+
+## Configuration
+
+- **[How to Configure Logging](configure-logging.md)** - Unify Kedro and Dagster logs and customize formatters.
+- **[How to Configure Custom Executors](configure-executors.md)** - Set up multiprocess, Docker, Kubernetes, Dask, or Celery executors.
+
+## Integration
+
+- **[How to Use MLflow](use-mlflow.md)** - Track experiments and artifacts with Kedro-MLflow and Dagster.
+- **[How to Use Dagster Partitions](use-partitions.md)** - Define partitioned datasets and map partitions across your pipeline.
+
+## Deployment
+
+- **[How to Deploy to Production](deploy-to-production.md)** - Move from local dev to a production Dagster deployment.
+- **[How to Migrate an Existing Kedro Project](migrate-existing-project.md)** - Add Dagster orchestration to an existing project.
+
+## Operations
+
+- **[Troubleshoot](troubleshoot.md)** - Diagnose and resolve common issues.
+
+## Contributing
+
+- **[Contributing to Kedro-Dagster](contribute.md)** - Development setup, workflow, and project standards.
