+++ docs/pages/how-to/troubleshoot.md
