+++ docs/pages/how-to/use-mlflow.md
@@ -24,3 +24,9 @@ Add an `mlflow.yml` to each environment's configuration directory (`conf/<ENV>/m
 ```
 
 See the [Example Project](../tutorials/example-project.md#mlflow-integration) for a complete multi-environment MLflow setup with Optuna integration.
+
+## See also
+
+- [Concepts](../explanation/concepts.md#mlflow-compatibility) - how hook preservation enables MLflow compatibility
+- [Architecture](../explanation/architecture.md#hook-preservation) - details on which hooks fire during Dagster execution
+- [How to Use Dagster Partitions](use-partitions.md) - note on hook limitations during backfills
