+++ docs/pages/reference/api.md
@@ -6,4 +6,14 @@ template: api-index.html
 
 Complete API reference for all Kedro-Dagster classes and functions. Use the search box to filter, or click any name to see full documentation.
 
+The API is organized by module:
+
+- **Translators** (`translator`, `catalog`, `nodes`, `pipelines`, `kedro`) - convert Kedro projects into Dagster definitions
+- **Creators** (`dagster`) - build Dagster executors, schedules, and loggers from configuration
+- **Datasets** (`datasets`) - custom dataset types for Dagster partitions and Nothing-typed assets
+- **Configuration** (`config`) - Pydantic models for `dagster.yml` sections (jobs, executors, schedules, loggers)
+- **Logging** (`logging`) - unified Kedro/Dagster logger and formatters
+- **CLI** (`cli`) - `kedro dagster` command implementations
+- **Utilities** (`utils`) - naming, formatting, and helper functions used internally
+
 <!-- API_TABLE -->
