+++ docs/pages/reference/datasets.md
@@ -1,0 +1,11 @@
+# Datasets
+
+Custom dataset implementations for Kedro-Dagster integration.
+
+## DagsterNothingDataset
+
+::: kedro_dagster.datasets.DagsterNothingDataset
+
+## DagsterPartitionedDataset
+
+::: kedro_dagster.datasets.DagsterPartitionedDataset
