+++ docs/pages/reference/index.md
@@ -1,0 +1,7 @@
+# Reference
+
+Technical descriptions of every configurable option, CLI command, and public API in Kedro-Dagster.
+
+- **[Configuration Reference](configuration.md)** - Complete field tables for `dagster.yml`: jobs, executors, schedules, and loggers.
+- **[CLI Reference](cli.md)** - All `kedro dagster` commands with flags, defaults, and examples.
+- **[API Reference](api.md)** - Auto-generated documentation for all public classes and functions.
