+++ docs/pages/tutorials/index.md
@@ -1,0 +1,15 @@
+# Tutorials
+
+Learn to use Kedro-Dagster through hands-on, step-by-step exercises. Start with Getting Started, then explore the Example Project for advanced patterns.
+
+<div class="grid cards" markdown>
+
+- **[Getting Started](getting-started.md)**
+
+    Install Kedro-Dagster and deploy your first Kedro pipeline to Dagster in minutes.
+
+- **[Example Project](example-project.md)**
+
+    Explore a complete project with dynamic pipelines, partitions, MLflow tracking, and multi-environment configuration.
+
+</div>
