+++ src/kedro_dagster/config/models.py
@@ -1,14 +1,253 @@
-"""Configuration definitions for Dagster executors.
+"""Pydantic configuration models for the Kedro-Dagster plugin.
 
-These pydantic models define the parameters accepted by Dagster executors when
-declared in ``dagster.yml`` under the ``executors`` section."""
+This module consolidates all configuration models used by the Kedro-Dagster
+plugin into a single file. Models are ordered bottom-up by dependency.
+"""
 
 from logging import getLogger
-from typing import Any
+from typing import TYPE_CHECKING, Any, Literal
 
-from pydantic import BaseModel, Field
+import yaml
+from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
+
+if TYPE_CHECKING:
+    from kedro.framework.context import KedroContext
 
 LOGGER = getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# Logging models
+# ---------------------------------------------------------------------------
+
+# Valid Python logging levels (normalized to uppercase)
+LogLevel = Literal["CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"]
+
+
+class LoggerOptions(BaseModel):
+    """Options for defining Dagster loggers.
+
+    Attributes
+    ----------
+    log_level : LogLevel
+        Logging level (CRITICAL/ERROR/WARNING/INFO/DEBUG/NOTSET).
+    handlers : list[dict[str, Any]]
+        List of handler config dicts.
+    formatters : dict[str, dict[str, Any]]
+        Formatter configs, name to config.
+    filters : dict[str, dict[str, Any]]
+        Filter configs, name to config.
+
+    See Also
+    --------
+    `kedro_dagster.dagster.LoggerCreator` :
+        Creates Dagster logger definitions from these options.
+    """
+
+    model_config = ConfigDict(validate_assignment=True, extra="forbid")
+
+    log_level: LogLevel = Field(
+        default="INFO",
+        description="Python logging level (case-insensitive)",
+    )
+    handlers: list[dict[str, Any]] = Field(
+        default=[],
+        description="List of handler configurations",
+    )
+    formatters: dict[str, dict[str, Any]] = Field(
+        default={},
+        description="Formatter configurations mapped by name",
+    )
+    filters: dict[str, dict[str, Any]] = Field(
+        default={},
+        description="Filter configurations mapped by name",
+    )
+
+    @field_validator("log_level", mode="before")
+    @classmethod
+    def normalize_log_level(cls, v: str) -> str:
+        """Normalize log level to uppercase for case-insensitive matching.
+
+        Parameters
+        ----------
+        v : str
+            Log level string to normalize.
+
+        Returns
+        -------
+        str
+            Normalized log level in uppercase.
+
+        Raises
+        ------
+        ValueError
+            If log level is not a valid string.
+        """
+        if not isinstance(v, str):
+            raise ValueError("Log level must be a string")
+
+        normalized = v.upper().strip()
+
+        # Check if the normalized level is valid
+        valid_levels = {"CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"}
+        if normalized not in valid_levels:
+            raise ValueError(
+                f"Invalid log level '{v}'. Must be one of: {', '.join(sorted(valid_levels))} (case-insensitive)"
+            )
+
+        return normalized
+
+    @field_validator("handlers")
+    @classmethod
+    def validate_handlers(cls, v: list[dict[str, Any]]) -> list[dict[str, Any]]:
+        """Validate handler configurations.
+
+        Parameters
+        ----------
+        v : list[dict[str, Any]]
+            Handler configurations to validate.
+
+        Returns
+        -------
+        list[dict[str, Any]]
+            Validated handler configurations.
+
+        Raises
+        ------
+        ValueError
+            If handler configuration is invalid.
+        """
+        for i, handler in enumerate(v):
+            if "class" not in handler:
+                raise ValueError(f"Handler at index {i} must specify a 'class' field")
+
+            if not isinstance(handler["class"], str):
+                raise ValueError(f"Handler class at index {i} must be a string")
+
+        return v
+
+    @field_validator("formatters")
+    @classmethod
+    def validate_formatters(cls, v: dict[str, dict[str, Any]]) -> dict[str, dict[str, Any]]:
+        """Validate formatter configurations.
+
+        Parameters
+        ----------
+        v : dict[str, dict[str, Any]]
+            Formatter configurations to validate.
+
+        Returns
+        -------
+        dict[str, dict[str, Any]]
+            Validated formatter configurations.
+
+        Raises
+        ------
+        ValueError
+            If formatter configuration is invalid.
+        """
+        for name, formatter in v.items():
+            # Require either 'format' for standard formatter or '()' for custom class
+            has_format = "format" in formatter
+            has_callable = "()" in formatter
+
+            if not (has_format or has_callable):
+                raise ValueError(f"Formatter '{name}' must specify either 'format' field or '()' for custom class")
+
+            # Type validation for keys when present
+            if has_format and not isinstance(formatter["format"], str):
+                raise ValueError(f"Formatter '{name}' 'format' must be a string")
+
+            if has_callable and not isinstance(formatter["()"], str):
+                raise ValueError(f"Formatter '{name}' '()' must be a string import path")
+
+        return v
+
+    @field_validator("filters")
+    @classmethod
+    def validate_filters(cls, v: dict[str, dict[str, Any]]) -> dict[str, dict[str, Any]]:
+        """Validate filter configurations.
+
+        Parameters
+        ----------
+        v : dict[str, dict[str, Any]]
+            Filter configurations to validate.
+
+        Returns
+        -------
+        dict[str, dict[str, Any]]
+            Validated filter configurations.
+
+        Raises
+        ------
+        ValueError
+            If filter configuration is invalid.
+        """
+        for name, filter_config in v.items():
+            # Require either custom callable/class path via '()' or a class path via 'class'
+            has_callable = "()" in filter_config
+            has_class = "class" in filter_config
+
+            if not (has_callable or has_class):
+                raise ValueError(
+                    f"Filter '{name}' must specify either '()' for custom callable/class or 'class' for import path"
+                )
+
+            # Basic type checks for keys when present
+            if has_callable and not isinstance(filter_config["()"], str):
+                raise ValueError(f"Filter '{name}' '()' must be a string import path")
+
+            if has_class and not isinstance(filter_config["class"], str):
+                raise ValueError(f"Filter '{name}' class must be a string")
+
+        return v
+
+    @model_validator(mode="after")
+    def validate_references(self) -> "LoggerOptions":
+        """Validate that handler/formatter/filter references are consistent.
+
+        Returns
+        -------
+        LoggerOptions
+            Self after validation.
+
+        Raises
+        ------
+        ValueError
+            If there are inconsistent references.
+        """
+        # Collect available formatter and filter names
+        available_formatters = set(self.formatters.keys()) if self.formatters else set()
+        available_filters = set(self.filters.keys()) if self.filters else set()
+
+        # Check handlers reference valid formatters/filters
+        if self.handlers:
+            for i, handler in enumerate(self.handlers):
+                # Check formatter reference
+                if "formatter" in handler:
+                    formatter_name = handler["formatter"]
+                    if formatter_name not in available_formatters:
+                        raise ValueError(
+                            f"Handler at index {i} references unknown formatter '{formatter_name}'. "
+                            f"Available formatters: {sorted(available_formatters)}"
+                        )
+
+                # Check filter references
+                if "filters" in handler:
+                    handler_filters = handler["filters"]
+                    if isinstance(handler_filters, list):
+                        for filter_name in handler_filters:
+                            if filter_name not in available_filters:
+                                raise ValueError(
+                                    f"Handler at index {i} references unknown filter '{filter_name}'. "
+                                    f"Available filters: {sorted(available_filters)}"
+                                )
+
+        return self
+
+
+# ---------------------------------------------------------------------------
+# Executor models
+# ---------------------------------------------------------------------------
 
 
 class InProcessExecutorOptions(BaseModel):
@@ -34,7 +273,7 @@ class InProcessExecutorOptions(BaseModel):
 
     See Also
     --------
-    `kedro_dagster.config.execution.MultiprocessExecutorOptions` :
+    `kedro_dagster.config.models.MultiprocessExecutorOptions` :
         Extends this class with concurrency settings.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -82,7 +321,7 @@ class MultiprocessExecutorOptions(InProcessExecutorOptions):
 
     See Also
     --------
-    `kedro_dagster.config.execution.InProcessExecutorOptions` :
+    `kedro_dagster.config.models.InProcessExecutorOptions` :
         Base class providing retry configuration.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -127,7 +366,7 @@ class DaskClusterConfig(BaseModel):
 
     See Also
     --------
-    `kedro_dagster.config.execution.DaskExecutorOptions` :
+    `kedro_dagster.config.models.DaskExecutorOptions` :
         Uses this as its cluster configuration.
     """
 
@@ -171,7 +410,7 @@ class DaskExecutorOptions(BaseModel):
 
     See Also
     --------
-    `kedro_dagster.config.execution.DaskClusterConfig` :
+    `kedro_dagster.config.models.DaskClusterConfig` :
         Cluster configuration consumed by this executor.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -220,7 +459,7 @@ class DockerExecutorOptions(MultiprocessExecutorOptions):
 
     See Also
     --------
-    `kedro_dagster.config.execution.MultiprocessExecutorOptions` :
+    `kedro_dagster.config.models.MultiprocessExecutorOptions` :
         Base class providing concurrency and retry settings.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -291,9 +530,9 @@ class CeleryExecutorOptions(BaseModel):
 
     See Also
     --------
-    `kedro_dagster.config.execution.CeleryDockerExecutorOptions` :
+    `kedro_dagster.config.models.CeleryDockerExecutorOptions` :
         Combines Celery with Docker container settings.
-    `kedro_dagster.config.execution.CeleryK8sJobExecutorOptions` :
+    `kedro_dagster.config.models.CeleryK8sJobExecutorOptions` :
         Combines Celery with Kubernetes job settings.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -342,9 +581,9 @@ class CeleryDockerExecutorOptions(CeleryExecutorOptions, DockerExecutorOptions):
 
     See Also
     --------
-    `kedro_dagster.config.execution.CeleryExecutorOptions` :
+    `kedro_dagster.config.models.CeleryExecutorOptions` :
         Provides Celery broker and backend configuration.
-    `kedro_dagster.config.execution.DockerExecutorOptions` :
+    `kedro_dagster.config.models.DockerExecutorOptions` :
         Provides Docker container configuration.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -403,7 +642,7 @@ class K8sJobConfig(BaseModel):
 
     See Also
     --------
-    `kedro_dagster.config.execution.K8sJobExecutorOptions` :
+    `kedro_dagster.config.models.K8sJobExecutorOptions` :
         Uses this as its step and per-step configuration.
     """
 
@@ -491,9 +730,9 @@ class K8sJobExecutorOptions(MultiprocessExecutorOptions):
 
     See Also
     --------
-    `kedro_dagster.config.execution.K8sJobConfig` :
+    `kedro_dagster.config.models.K8sJobConfig` :
         Kubernetes job configuration consumed by this executor.
-    `kedro_dagster.config.execution.CeleryK8sJobExecutorOptions` :
+    `kedro_dagster.config.models.CeleryK8sJobExecutorOptions` :
         Combines Celery with Kubernetes job settings.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -603,9 +842,9 @@ class CeleryK8sJobExecutorOptions(CeleryExecutorOptions, K8sJobExecutorOptions):
 
     See Also
     --------
-    `kedro_dagster.config.execution.CeleryExecutorOptions` :
+    `kedro_dagster.config.models.CeleryExecutorOptions` :
         Provides Celery broker and backend configuration.
-    `kedro_dagster.config.execution.K8sJobExecutorOptions` :
+    `kedro_dagster.config.models.K8sJobExecutorOptions` :
         Provides Kubernetes job configuration.
     `kedro_dagster.dagster.ExecutorCreator` :
         Builds Dagster executor definitions from these options.
@@ -642,3 +881,290 @@ EXECUTOR_MAP = {
     "celery_docker_executor": CeleryDockerExecutorOptions,
     "celery_k8s_job_executor": CeleryK8sJobExecutorOptions,
 }
+
+
+# ---------------------------------------------------------------------------
+# Schedule models
+# ---------------------------------------------------------------------------
+
+
+class ScheduleOptions(BaseModel):
+    """Options for defining Dagster schedules.
+
+    Attributes
+    ----------
+    cron_schedule : str
+        Cron expression for the schedule.
+    execution_timezone : str or None
+        Timezone in which the schedule should execute.
+    description : str or None
+        Optional description of the schedule.
+    metadata : dict[str, Any] or None
+        Additional metadata for the schedule.
+
+    Examples
+    --------
+    ```yaml
+    schedules:
+        daily_schedule:
+            cron_schedule: "0 6 * * *"
+            execution_timezone: "UTC"
+            description: "Run every morning"
+            metadata:
+                owner: data-platform
+    ```
+
+    See Also
+    --------
+    `kedro_dagster.dagster.ScheduleCreator.create_schedules` :
+        Builds Dagster schedule definitions from these options.
+    """
+
+    cron_schedule: str
+    execution_timezone: str | None = None
+    description: str | None = None
+    metadata: dict[str, Any] | None = None
+
+
+# ---------------------------------------------------------------------------
+# Pipeline and Job models
+# ---------------------------------------------------------------------------
+
+
+class PipelineOptions(BaseModel):
+    """Options for filtering and configuring Kedro pipelines within a Dagster job.
+
+    Attributes
+    ----------
+    pipeline_name : str
+        Name of the Kedro pipeline to run. Defaults to ``__default__``.
+    from_nodes : list[str] or None
+        List of node names to start execution from.
+    to_nodes : list[str] or None
+        List of node names to end execution at.
+    node_names : list[str] or None
+        List of specific node names to include in the pipeline.
+    from_inputs : list[str] or None
+        List of dataset names to use as entry points.
+    to_outputs : list[str] or None
+        List of dataset names to use as exit points.
+    node_namespaces : list[str] or None
+        Namespace(s) to filter nodes by.
+    tags : list[str] or None
+        List of tags to filter nodes by.
+
+    Examples
+    --------
+    ```yaml
+    jobs:
+      sales_etl:
+        pipeline:
+          pipeline_name: etl
+          node_namespaces: ["sales", "shared"]
+          tags: ["daily", "priority"]
+          from_nodes: ["extract_raw_sales"]
+          to_nodes: ["publish_clean_sales"]
+          from_inputs: ["raw_sales"]
+          to_outputs: ["clean_sales"]
+    ```
+
+    See Also
+    --------
+    `kedro_dagster.config.models.JobOptions` :
+        Wraps this model alongside executor and schedule settings.
+    `kedro_dagster.utils.get_filter_params_dict` :
+        Extracts filter parameters from this configuration.
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    pipeline_name: str = "__default__"
+    from_nodes: list[str] | None = None
+    to_nodes: list[str] | None = None
+    node_names: list[str] | None = None
+    from_inputs: list[str] | None = None
+    to_outputs: list[str] | None = None
+    node_namespaces: list[str] | None = None
+    tags: list[str] | None = None
+
+
+class JobOptions(BaseModel):
+    """Configuration options for a Dagster job.
+
+    Attributes
+    ----------
+    pipeline : PipelineOptions
+        Pipeline options specifying which pipeline and nodes to run.
+    executor : ExecutorOptions or str or None
+        Executor options instance or string key referencing an executor.
+    schedule : ScheduleOptions or str or None
+        Schedule options instance or string key referencing a schedule.
+    loggers : list[LoggerOptions or str] or None
+        List of logger configurations (inline ``LoggerOptions``) or
+        logger names (strings) to attach to the job.
+
+    Examples
+    --------
+    ```yaml
+    jobs:
+      my_data_processing:
+        pipeline:
+          pipeline_name: data_processing
+          node_namespaces: [price_predictor]
+          tags: [test1]
+        executor: multiprocessing
+        schedule: daily_schedule
+        loggers: [console, file_logger]
+    ```
+
+    See Also
+    --------
+    `kedro_dagster.config.models.PipelineOptions` :
+        Pipeline filtering options within this job.
+    `kedro_dagster.config.models.ScheduleOptions` :
+        Schedule options referenced by this job.
+    `kedro_dagster.config.models.LoggerOptions` :
+        Logger options referenced by this job.
+    `kedro_dagster.config.models.KedroDagsterConfig` :
+        Top-level config that holds a dict of these job options.
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    pipeline: PipelineOptions
+    executor: ExecutorOptions | str | None = None
+    schedule: ScheduleOptions | str | None = None
+    loggers: list[LoggerOptions | str] | None = None
+
+
+# ---------------------------------------------------------------------------
+# Top-level config
+# ---------------------------------------------------------------------------
+
+
+class KedroDagsterConfig(BaseModel):
+    """Main configuration class representing the ``dagster.yml`` structure.
+
+    Attributes
+    ----------
+    loggers : dict[str, LoggerOptions] or None
+        Mapping of logger names to logger options.
+    executors : dict[str, ExecutorOptions] or None
+        Mapping of executor names to executor options.
+    schedules : dict[str, ScheduleOptions] or None
+        Mapping of schedule names to schedule options.
+    jobs : dict[str, JobOptions] or None
+        Mapping of job names to job options.
+
+    See Also
+    --------
+    `kedro_dagster.config.models.get_dagster_config` :
+        Loads and returns an instance of this class.
+    """
+
+    model_config = ConfigDict(validate_assignment=True, extra="forbid")
+
+    loggers: dict[str, LoggerOptions] | None = None
+    executors: dict[str, ExecutorOptions] | None = None
+    schedules: dict[str, ScheduleOptions] | None = None
+    jobs: dict[str, JobOptions] | None = None
+
+    @model_validator(mode="before")
+    @classmethod
+    def validate_executors(cls, values: dict[str, Any]) -> dict[str, Any]:
+        """Parse raw executor config dicts into typed executor option models.
+
+        Parameters
+        ----------
+        values : dict[str, Any]
+            Raw model data before validation.
+
+        Returns
+        -------
+        dict[str, Any]
+            Model data with executor configs replaced by option instances.
+
+        Raises
+        ------
+        ValueError
+            If an executor type is not recognized.
+        """
+        executors = values.get("executors") or {}
+
+        parsed_executors = {}
+        for name, executor_config in executors.items():
+            if "in_process" in executor_config:
+                executor_name = "in_process"
+            elif "multiprocess" in executor_config:
+                executor_name = "multiprocess"
+            elif "k8s_job_executor" in executor_config:
+                executor_name = "k8s_job_executor"
+            elif "docker_executor" in executor_config:
+                executor_name = "docker_executor"
+            else:
+                msg = f"Unknown executor type in {name}"
+                LOGGER.error(msg)
+                raise ValueError(msg)
+
+            executor_options_class = EXECUTOR_MAP[executor_name]
+            executor_options_params = executor_config[executor_name] or {}
+            parsed_executors[name] = executor_options_class(**executor_options_params)
+
+        values["executors"] = parsed_executors
+        return values
+
+
+def get_dagster_config(context: "KedroContext") -> "KedroDagsterConfig":
+    """Get the Dagster configuration from the ``dagster.yml`` file.
+
+    Parameters
+    ----------
+    context : KedroContext
+        ``KedroContext`` that was created.
+
+    Returns
+    -------
+    KedroDagsterConfig
+        Dagster configuration.
+
+    See Also
+    --------
+    `kedro_dagster.config.models.KedroDagsterConfig` :
+        The configuration model returned.
+    """
+    from kedro.config import MissingConfigException
+
+    LOGGER.info("Loading Dagster configuration...")
+
+    try:
+        if "dagster" not in context.config_loader.config_patterns:
+            context.config_loader.config_patterns.update({"dagster": ["dagster*", "dagster*/**", "**/dagster*"]})
+        conf_dagster_yml = context.config_loader["dagster"]
+    except MissingConfigException:
+        LOGGER.warning(
+            "No 'dagster.yml' config file found in environment. Default configuration will be used. "
+            "Use ``kedro dagster init`` command in CLI to customize the configuration."
+        )
+
+        conf_dagster_yml = {}
+
+    dagster_config = KedroDagsterConfig(**conf_dagster_yml)
+
+    return dagster_config
+
+
+# ---------------------------------------------------------------------------
+# Config template
+# ---------------------------------------------------------------------------
+
+CONFIG_TEMPLATE_YAML: str = """\
+# dagster.yml - Kedro-Dagster orchestration configuration
+# See https://kedro-dagster.readthedocs.io/en/latest/pages/reference/configuration/
+
+jobs:
+  __default__:
+    pipeline:
+      pipeline_name: __default__
+"""
+
+_CONFIG_TEMPLATE: KedroDagsterConfig = KedroDagsterConfig.model_validate(yaml.safe_load(CONFIG_TEMPLATE_YAML))
